Backened/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.4.4
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
PyMySQL==1.0.3
python-dotenv==1.0.0
